Create RubyClaimEvidenceApi Class
